define constant objects for style class names
